Insert recurly data into shared ext table
